# marketing-contracts/src/main/resources/openapi/marketing-api.yaml
openapi: 3.1.0
info:
  title: Marketing Low-code Platform API
  version: 1.0.0
  description: R1 edge contract. Tenant and organization scope come from the verified identity in OIDC mode.
servers:
  - {url: http://localhost:8080, description: Local edge gateway}
security: [{oidc: []}]
...
paths:
  /api/v1/campaigns:
    get: {tags: [Control], operationId: listCampaigns, responses: *list}
    post:
      tags: [Control]
      operationId: createCampaign
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/CreateCampaign'}
      responses: *created
  /api/v1/definitions:
    post:
      tags: [Control]
      operationId: saveDefinition
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/SaveDefinition'}
      responses: *created
  /api/v1/definitions/latest:
    get:
      tags: [Control]
      operationId: getLatestDefinition
      parameters:
        - {in: query, name: campaignId, required: true, schema: {type: string}}
        - {in: query, name: dialect, required: true, schema: {type: string, enum: [OFFER_DECISION_DAG, AUDIENCE_EXPRESSION, JOURNEY_STATE_MACHINE, DMN_DECISION_TABLE, BENEFIT_POLICY]}}
      responses: *ok
  /api/v1/definitions/{definitionId}/versions/{version}:
    get:
      tags: [Control]
      operationId: getDefinitionVersion
      parameters: [{$ref: '#/components/parameters/DefinitionId'}, {$ref: '#/components/parameters/Version'}]
      responses: *ok
  /api/v1/definitions/{definitionId}/versions/{version}:validate:
    post:
      tags: [Control]
      operationId: validateDefinition
      parameters: [{$ref: '#/components/parameters/DefinitionId'}, {$ref: '#/components/parameters/Version'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: *ok
  /api/v1/definitions/{definitionId}/versions/{version}:simulate:
    post:
      tags: [Control]
      operationId: simulateDefinition
      parameters: [{$ref: '#/components/parameters/DefinitionId'}, {$ref: '#/components/parameters/Version'}]
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *ok
  /api/v1/definitions/{definitionId}/versions/{version}:submit:
    post:
      tags: [Control]
      operationId: submitDefinition
      parameters: [{$ref: '#/components/parameters/DefinitionId'}, {$ref: '#/components/parameters/Version'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: *ok
  /api/v1/definitions/{definitionId}/versions/{version}/terms:
    get:
      tags: [Control]
      operationId: getDefinitionTerms
      parameters: [{$ref: '#/components/parameters/DefinitionId'}, {$ref: '#/components/parameters/Version'}]
      responses: *ok
  /api/v1/approvals:
    get: {tags: [Control], operationId: listApprovals, responses: *list}
  /api/v1/approvals/{caseId}/decisions:
    post:
      tags: [Control]
      operationId: decideApproval
      parameters: [{$ref: '#/components/parameters/CaseId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/ApprovalDecision'}
      responses: *ok
  /api/v1/registries/nodes:
    get: {tags: [Control], operationId: listNodeDefinitions, responses: *list}
  /api/v1/releases:
    get: {tags: [Release], operationId: listReleases, responses: *list}
    post:
      tags: [Release]
      operationId: stageRelease
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/StageRelease'}
      responses: *created
  /api/v1/releases/{manifestId}:ack:
    post:
      tags: [Release]
      operationId: acknowledgeRelease
      parameters: [{$ref: '#/components/parameters/ManifestId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *ok
  /api/v1/releases/{manifestId}:activate:
    post:
      tags: [Release]
      operationId: activateRelease
      parameters: [{$ref: '#/components/parameters/ManifestId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: *ok
  /api/v1/releases/{manifestId}:rollback:
    post:
      tags: [Release]
      operationId: rollbackRelease
      parameters: [{$ref: '#/components/parameters/ManifestId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *ok
  /api/v1/releases/desired:
    get:
      tags: [Release]
      operationId: getDesiredRelease
      parameters:
        - {in: query, name: environment, required: true, schema: {type: string}}
        - {in: query, name: cell, required: true, schema: {type: string}}
        - {in: query, name: runtime, required: true, schema: {type: string}}
        - {in: query, name: namespace, required: true, schema: {type: string}}
      responses: *ok
  /api/v1/releases/kill-switches/{namespace}:
    put:
      tags: [Release]
      operationId: setKillSwitch
      parameters: [{in: path, name: namespace, required: true, schema: {type: string}}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *ok
...
  /api/v1/journey-runtime/manifest:
    put: {tags: [Journey], operationId: warmJourneyRuntime, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/journey-runtime/activation:
    put: {tags: [Journey], operationId: activateJourneyRuntime, requestBody: {$ref: '#/components/requestBodies/JsonObject'}, responses: *ok}
  /api/v1/journeys/{journeyId}/versions/{version}:
    put:
      tags: [Journey]
      operationId: registerJourneyForDevelopment
      description: DEV only; production installs a signed artifact through the runtime manifest endpoint.
      parameters: [{$ref: '#/components/parameters/JourneyId'}, {$ref: '#/components/parameters/Version'}]
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *ok
  /api/v1/enrollments:
    get:
      tags: [Journey]
      operationId: listJourneyEnrollments
      parameters:
        - {in: query, name: status, required: false, schema: {type: string, enum: [RUNNING, WAITING, COMPLETED, FAILED]}}
        - {in: query, name: journeyId, required: false, schema: {type: string}}
        - {in: query, name: limit, required: false, schema: {type: integer, default: 50, minimum: 1, maximum: 200}}
      responses: *list
    post:
      tags: [Journey]
      operationId: enrollJourneyForDevelopment
      description: DIRECT/DEV only; production submits a JOURNEY_SIGNAL START envelope to POST /api/v1/events.
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *created
  /api/v1/enrollments/{enrollmentId}/signals:
    post:
      tags: [Journey]
      operationId: signalJourneyForDevelopment
      description: DIRECT/DEV only; production uses POST /api/v1/events so Flink is the only state writer.
      parameters: [{$ref: '#/components/parameters/EnrollmentId'}]
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *ok
  /api/v1/enrollments/{enrollmentId}:
    get: {tags: [Journey], operationId: getJourneyEnrollment, parameters: [{$ref: '#/components/parameters/EnrollmentId'}], responses: *ok}
  /api/v1/journeys/{journeyId}:migrate:
    post:
      tags: [Journey]
      operationId: migrateJourneyForDevelopment
      description: DIRECT/DEV only; production stream state stays version-pinned and is migrated with savepoints.
      parameters: [{$ref: '#/components/parameters/JourneyId'}]
      requestBody: {$ref: '#/components/requestBodies/JsonObject'}
      responses: *ok
...
components:
  securitySchemes:
    oidc: {type: openIdConnect, openIdConnectUrl: https://identity.example.invalid/.well-known/openid-configuration}
  parameters:
    IdempotencyKey: {in: header, name: Idempotency-Key, required: true, schema: {type: string, minLength: 1, maxLength: 128}}
    DefinitionId: {in: path, name: definitionId, required: true, schema: {type: string}}
    Version: {in: path, name: version, required: true, schema: {type: integer, format: int64, minimum: 1}}
...
    CaseId: {in: path, name: caseId, required: true, schema: {type: string}}
    ManifestId: {in: path, name: manifestId, required: true, schema: {type: string}}
...
    JourneyId: {in: path, name: journeyId, required: true, schema: {type: string}}
    EnrollmentId: {in: path, name: enrollmentId, required: true, schema: {type: string}}
...
  requestBodies:
    JsonObject: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/JsonObject'}}}}
...
    CreateCampaign: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateCampaign'}}}}
    SaveDefinition: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SaveDefinition'}}}}
    ApprovalDecision: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalDecision'}}}}
    StageRelease: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StageReleaseRequest'}}}}
...
  schemas:
    JsonObject: {type: object, additionalProperties: true}
    CreateCampaign:
      type: object
      additionalProperties: false
      required: [name, objective, organizationId]
      properties: {name: {type: string, minLength: 1}, objective: {type: string, minLength: 1}, organizationId: {type: string, minLength: 1}, shopId: {type: [string, 'null']}}
    SaveDefinition:
      type: object
      additionalProperties: false
      required: [campaignId, graph]
      properties: {campaignId: {type: string, minLength: 1}, graph: {$ref: '../schema/graph/graph-definition.schema.json'}}
    ApprovalDecision:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role: {type: string, enum: [BUSINESS, FINANCE, COMPLIANCE, MERCHANT]}
        decision: {type: string, enum: [APPROVE, REJECT], default: APPROVE}
        comment: {type: string, maxLength: 128}
    StageReleaseRequest:
      type: object
      additionalProperties: false
      required: [definitionId, definitionVersion, environment, cell, runtime, namespace, artifacts, schemaVersions, canaryBasisPoints, activationAt, approvalCaseIds]
      properties:
        definitionId: {type: string}
        definitionVersion: {type: integer, format: int64, minimum: 1}
        environment: {type: string}
        cell: {type: string}
        runtime: {type: string}
        namespace: {type: string}
        artifacts: {type: array, items: {$ref: '#/components/schemas/JsonObject'}}
        schemaVersions: {type: object, additionalProperties: {type: string}}
        canaryBasisPoints: {type: integer, minimum: 0, maximum: 10000}
        activationAt: {type: string, format: date-time}
        approvalCaseIds: {type: array, items: {type: string}}
...